pyproject.toml
[project]
name = "aistation"
version = "0.1.0"
description = "Unofficial Python SDK for Westlake University AI Station"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }
authors = [{ name = "wzq" }]
dependencies = [
    "requests>=2.32",
    "gmssl>=3.2",
    "pyyaml>=6.0",
    "platformdirs>=4.0",
]

[project.optional-dependencies]
cli = [
    "typer>=0.12",
    "rich>=13",
    "pillow>=10",
]
dev = [
    "pytest>=8",
    "pytest-cov",
    "pytest-mock",
    "ruff",
    "mypy",
    "responses>=0.25",
]

[project.scripts]
aistation = "aistation.cli.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aistation"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.mypy]
python_version = "3.12"
strict = true
warn_unused_ignores = true
